# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Daniel Ernesto Lopez Sibrian",
        "description": "API de ejemplo para la clase de desarrollo de backend y apis",
        "contact": {
            "name": "Daniel Ernesto Lopez Sibrian",
            "url": "https://localhost/api/",
            "email": "ia.daniellopez1@ufg.edu.sv"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/obtener": {
            "get": {
                "tags": [
                    "obtener"
                ],
                "summary": "Esta es un endpoint/metodo para obtener datos y mostrarlos al usuario",
                "description": "Descripcion de prueba para la clase de desarrollo de backend y api desarrollada de la clase 19/10",
                "operationId": "2572d9320dfbb5aeaf9a1d7d6e238a2c",
                "responses": {
                    "200": {
                        "description": "Se devuelven todos los registros cuando la consulta sea satisfactoria"
                    },
                    "500": {
                        "description": "No se pudo realizar la ejecución al parecer hubo un error interno"
                    }
                }
            }
        },
        "/api/nuevo": {
            "post": {
                "tags": [
                    "nuevo"
                ],
                "summary": "Esta es un endpoint/metodo para obtener datos y mostrarlos al usuario",
                "description": "Descripcion de prueba para la clase de desarrollo de backend y api desarrollada de la clase 19/10",
                "operationId": "dd566d3abe549d9562cd9048b0b0f8f6",
                "responses": {
                    "200": {
                        "description": "Se devuelven todos los registros cuando la consulta sea satisfactoria"
                    },
                    "500": {
                        "description": "No se pudo realizar la ejecución al parecer hubo un error interno"
                    }
                }
            }
        },
        "/api/ejecutar": {
            "get": {
                "tags": [
                    "ejecutar"
                ],
                "summary": "Esta es un endpoint/metodo para obtener datos y mostrarlos al usuario",
                "description": "Descripcion de prueba para la clase de desarrollo de backend y api desarrollada de la clase 19/10",
                "operationId": "ed42aafacf51d9aa570647800fcaa0e9",
                "responses": {
                    "200": {
                        "description": "Se devuelven todos los registros cuando la consulta sea satisfactoria"
                    },
                    "500": {
                        "description": "No se pudo realizar la ejecución al parecer hubo un error interno"
                    }
                }
            }
        }
    }
}